lxd/device: Only remove unix socket file when `bind=host`
Fixes https://github.com/canonical/lxd/issues/15782.

+++ lxd/device/proxy.go
@@ -687,7 +687,7 @@ func (d *proxy) killProxyProc(pidPath string) error {
 
 	// Remove socket file if needed.
 	// Unix non-abstract sockets are addressed to the host filesystem, not scoped inside the LXD snap.
-	if listenAddr.ConnType == "unix" && !listenAddr.Abstract {
+	if listenAddr.ConnType == "unix" && !listenAddr.Abstract && d.config["bind"] == "host" {
 		err = os.Remove(shared.HostPath(listenAddr.Address))
 		if err != nil && !errors.Is(err, os.ErrNotExist) {
 			return fmt.Errorf("Failed to remove socket file: %w", err)
